+++ src/components/Header/DesktopView/DesktopView.jsx
@@ -1,0 +1,28 @@
+import {  MainDiv, MainDivLoggedIn, NavIconsDiv } from './DesktopViewStyled';
+import { Logo } from '../Logo/Logo';
+import { UserNav } from '../UserNav/UserNav';
+import { UserBar } from '../UserBar/UserBar';
+
+export const DesktopView = () => {
+
+const isLoggedIn = true;
+
+  return (
+    <>
+    {
+      isLoggedIn?
+     <MainDivLoggedIn>
+        <Logo/>
+        <NavIconsDiv> 
+              <UserNav/>
+             <UserBar/>
+        </NavIconsDiv>
+  </MainDivLoggedIn > :
+
+          <MainDiv>
+            <Logo/>
+          </MainDiv>
+    }
+    </> )
+   
+}
