+++ src/components/Header/LogOut/LogOutStyled.jsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Text = styled.p`   
+    color: #EFEDE8;
+    font-family: Roboto;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+   line-height: 1.28;
+
+    @media (min-width: 768px) {
+    font-size: 16px;
+    line-height: 1.5; 
+  }
+`
