+++ src/components/Header/Logo/Logo.jsx
@@ -1,0 +1,21 @@
+import useMediaQuery from '@mui/material/useMediaQuery';
+
+import { Link } from 'react-router-dom';
+import sprite from "../../../assets/sprite.svg"
+
+
+export const Logo = () => {
+
+     const isLogedIn = true;
+     const isMobileScreen = useMediaQuery('(max-width: 767px)');
+
+
+  return (
+    <Link to="/">
+         <svg width={isMobileScreen ? "82px" : "151px"} height={isMobileScreen ? "12px" : "17px"}>
+                  <use xlinkHref={sprite + "#icon-Logo-1" } />
+              </svg>
+      </Link>
+     )
+ 
+};
