--- src/components/Header/header.css
@@ -1,9 +1,0 @@
-.nav {
-    display: flex;
-    flex-direction: row;
-    row-gap: 14px;
-    justify-content: space-around;
-    align-items: center;
-    background-color: transparent;
-    height: 60px;
-}
